+++ lib_sw_pll/src/sw_pll_pfd.h
@@ -1,11 +1,9 @@
-// Copyright 2023 XMOS LIMITED.
+// Copyright 2023-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 #include <stdint.h>
 #include <stddef.h>
-#ifndef __XC__
-#include <xcore/assert.h>
-#endif
+
 #include "sw_pll_common.h"
 
 #pragma once